# Conflicts:
#	meetingManager/src/main/scala/videomeeting/meetingManager/core/MeetingActor.scala

+++ pcClient/src/main/scala/videomeeting/pcClient/scene/StartScene.scala
@@ -72,7 +72,7 @@ object StartScene {
 
     def stopLive()
 
-    def audienceAcceptance(userId: Long, accept: Boolean, newRequest: SpeakListInfo)
+    def audienceAcceptance(userId: Int, accept: Boolean, newRequest: SpeakListInfo)
 
     def shutJoin()
 
@@ -257,7 +257,7 @@ class StartScene(stage: Stage) {
     * 更新连线请求
     *
     **/
-  def updateAudienceList(audienceId: Long, audienceName: String): Unit = {
+  def updateAudienceList(audienceId: Int, audienceName: String): Unit = {
     if (!tb2.isSelected) {
       tb2.setGraphic(applyIcon1)
     }
